Refactored operatioal area approval into operationalareaapproval
