+++ govapp/apps/operationalarea/apps.py
@@ -1,6 +1,6 @@
 from django.apps import AppConfig
 
 
-class ApprovalConfig(AppConfig):
+class OperationalAreaConfig(AppConfig):
     default_auto_field = "django.db.models.BigAutoField"
-    name = "govapp.apps.approval"
+    name = "govapp.apps.operationalarea"
