+++ govapp/apps/operationalarea/models.py
@@ -4,13 +4,48 @@ from django.contrib.gis.db.models import MultiLineStringField, MultiPolygonField
 from django.db import models
 from model_utils import Choices
 from model_utils.models import StatusModel, TimeStampedModel
+from protected_media.models import ProtectedFileField
 
-from govapp.apps.approval.models import Approval
 from govapp.apps.burnplanning.models import BurnPlanUnit
-from govapp.apps.main.models import ReferenceableModel, UniqueNameableModel
+from govapp.apps.main.models import (
+    Lga,
+    ReferenceableModel,
+    UniqueNameableModel,
+    file_upload_location,
+)
 from govapp.apps.risk.models import RiskFactor
 
 logger = getLogger(__name__)
+
+
+class OperationalAreaApproval(TimeStampedModel):
+    """Burn Program and Operational Area approvals"""
+
+    # TODO What is the difference between an approval and an endorsement?
+    APPROVAL_TYPE = Choices(
+        ("approval", "Approval"),
+        ("endorsement", "Endorsement"),
+    )
+    approver: models.CharField = models.CharField(
+        max_length=255, null=True, blank=True
+    )  # Corporate Executive, Shire, Other Lands, Owner
+    lga: models.ForeignKey = models.ForeignKey(
+        Lga, on_delete=models.PROTECT, null=True, blank=True
+    )  # Shire
+
+    can_provide_evidence: models.BooleanField = models.BooleanField(
+        default=False
+    )  # Whether the user can attach files, texts, or remove the approval
+    file_as_approval = ProtectedFileField(upload_to=file_upload_location)
+    text_as_approval: models.TextField = models.TextField(null=True, blank=True)
+    text_remove_justification: models.TextField = models.TextField(
+        null=True, blank=True
+    )
+
+    def __str__(self):
+        if self.lga:
+            return f"{self.approver} {self.APPROVAL_TYPE} {self.lga}"
+        return f"{self.approver} {self.APPROVAL_TYPE}"
 
 
 class OperationalArea(
@@ -49,7 +84,7 @@ class OperationalArea(
 
     # Legal / Approvals
     approvals: models.ManyToManyField = models.ManyToManyField(
-        Approval, blank=True, related_name="operational_areas"
+        OperationalAreaApproval, blank=True, related_name="operational_areas"
     )
 
     # Risk Factors
